Moved communication implementation to common hal sources
