+++ firmware/src/hal/src/ledger/communication.c
@@ -1,1 +1,1 @@
-../x86/communication.c
+../common/communication.c
